+++ tests/FakeRequest.php
@@ -20,6 +20,7 @@ final class FakeRequest implements ServerRequestInterface
      * @param array<string, string|int|list<string>> $queryParams
      * @param array<string, list<string>> $headers
      * @param array<array-key, mixed>|object|null $parsedBody
+     * @param array<string, mixed> $attributes
      */
     public function __construct(
         private readonly string $method = 'POST',
@@ -30,6 +31,7 @@ final class FakeRequest implements ServerRequestInterface
         private readonly array $serverParams = [],
         private readonly array $queryParams = [],
         private readonly array|object|null $parsedBody = null,
+        private readonly array $attributes = [],
     ) {}
 
     #[\Override]
@@ -131,13 +133,13 @@ final class FakeRequest implements ServerRequestInterface
     #[\Override]
     public function getAttributes(): array
     {
-        return [];
+        return $this->attributes;
     }
 
     #[\Override]
-    public function getAttribute(string $name, $default = null): null
+    public function getAttribute(string $name, mixed $default = null): mixed
     {
-        return null;
+        return $this->attributes[$name] ?? $default;
     }
 
     #[\Override]